move maintenance:* -> apps:maintenance:*
