+++ index.js
@@ -12,7 +12,7 @@ exports.topics = [
   { name: 'info',          hidden: true },
   { name: 'kill',          hidden: true },
   { name: 'labs',          description: 'experimental features' },
-  { name: 'maintenance',   description: 'manage maintenance mode for an app' },
+  { name: 'maintenance',   hidden: true },
   { name: 'notifications', description: 'display notifications' },
   { name: 'open',          hidden: true },
   { name: 'ps',            description: 'manage dynos (dynos, workers)' },
@@ -39,6 +39,9 @@ exports.commands = [
   require('./commands/apps/favorites/remove'),
   require('./commands/apps/info').apps,
   require('./commands/apps/info').root,
+  require('./commands/apps/maintenance'),
+  require('./commands/apps/maintenance/off'),
+  require('./commands/apps/maintenance/on'),
   require('./commands/apps/open').open,
   require('./commands/apps/open').root,
   require('./commands/apps/rename').apps,
@@ -78,9 +81,6 @@ exports.commands = [
   require('./commands/labs/disable'),
   require('./commands/labs/enable'),
   require('./commands/labs/info'),
-  require('./commands/maintenance'),
-  require('./commands/maintenance/off'),
-  require('./commands/maintenance/on'),
   require('./commands/notifications'),
   require('./commands/ps'),
   require('./commands/ps/restart').dyno,
